refactor(whatsapp): move normalize + resolve-outbound-target to extensions
Move 4 remaining src/whatsapp/ files to extensions/whatsapp/src/ and
rewire all 11 importers. Adds isWhatsAppUserTarget export (new in
upstream). src/whatsapp/ directory is now empty.

Partial cherry-pick of upstream commit 2131981230 (11 of 143 files)
plus deletions from 042669d8c8.

Cherry-picked-from: openclaw/openclaw@2131981230
Cherry-picked-from: openclaw/openclaw@042669d8c8
Closes #2036
