+++ extensions/whatsapp/src/normalize-target.test.ts
@@ -1,5 +1,9 @@
 import { describe, expect, it } from "vitest";
-import { isWhatsAppGroupJid, isWhatsAppUserTarget, normalizeWhatsAppTarget } from "./normalize.js";
+import {
+  isWhatsAppGroupJid,
+  isWhatsAppUserTarget,
+  normalizeWhatsAppTarget,
+} from "./normalize-target.js";
 
 describe("normalizeWhatsAppTarget", () => {
   it("preserves group JIDs", () => {
@@ -15,11 +19,8 @@ describe("normalizeWhatsAppTarget", () => {
   });
 
   it("normalizes user JIDs with device suffix to E.164", () => {
-    // This is the bug fix: JIDs like "41796666864:0@s.whatsapp.net" should
-    // normalize to "+41796666864", not "+417966668640" (extra digit from ":0")
     expect(normalizeWhatsAppTarget("41796666864:0@s.whatsapp.net")).toBe("+41796666864");
     expect(normalizeWhatsAppTarget("1234567890:123@s.whatsapp.net")).toBe("+1234567890");
-    // Without device suffix still works
     expect(normalizeWhatsAppTarget("41796666864@s.whatsapp.net")).toBe("+41796666864");
   });
