+++ extensions/whatsapp/src/normalize-target.ts
@@ -1,4 +1,4 @@
-import { normalizeE164 } from "../utils.js";
+import { normalizeE164 } from "../../../src/utils.js";
 
 const WHATSAPP_USER_JID_RE = /^(\d+)(?::\d+)?@s\.whatsapp\.net$/i;
 const WHATSAPP_LID_RE = /^(\d+)@lid$/i;
@@ -27,19 +27,11 @@ export function isWhatsAppGroupJid(value: string): boolean {
   return /^[0-9]+(-[0-9]+)*$/.test(localPart);
 }
 
-/**
- * Check if value looks like a WhatsApp user target (e.g. "41796666864:0@s.whatsapp.net" or "123@lid").
- */
 export function isWhatsAppUserTarget(value: string): boolean {
   const candidate = stripWhatsAppTargetPrefixes(value);
   return WHATSAPP_USER_JID_RE.test(candidate) || WHATSAPP_LID_RE.test(candidate);
 }
 
-/**
- * Extract the phone number from a WhatsApp user JID.
- * "41796666864:0@s.whatsapp.net" -> "41796666864"
- * "123456@lid" -> "123456"
- */
 function extractUserJidPhone(jid: string): string | null {
   const userMatch = jid.match(WHATSAPP_USER_JID_RE);
   if (userMatch) {
@@ -61,7 +53,6 @@ export function normalizeWhatsAppTarget(value: string): string | null {
     const localPart = candidate.slice(0, candidate.length - "@g.us".length);
     return `${localPart}@g.us`;
   }
-  // Handle user JIDs (e.g. "41796666864:0@s.whatsapp.net")
   if (isWhatsAppUserTarget(candidate)) {
     const phone = extractUserJidPhone(candidate);
     if (!phone) {
@@ -70,8 +61,6 @@ export function normalizeWhatsAppTarget(value: string): string | null {
     const normalized = normalizeE164(phone);
     return normalized.length > 1 ? normalized : null;
   }
-  // If the caller passed a JID-ish string that we don't understand, fail fast.
-  // Otherwise normalizeE164 would happily treat "group:120@g.us" as a phone number.
   if (candidate.includes("@")) {
     return null;
   }
