+++ extensions/whatsapp/src/resolve-outbound-target.ts
@@ -1,5 +1,5 @@
-import { missingTargetError } from "../infra/outbound/target-errors.js";
-import { isWhatsAppGroupJid, normalizeWhatsAppTarget } from "./normalize.js";
+import { missingTargetError } from "../../../src/infra/outbound/target-errors.js";
+import { isWhatsAppGroupJid, normalizeWhatsAppTarget } from "./normalize-target.js";
 
 export type WhatsAppOutboundTargetResolution =
   | { ok: true; to: string }
@@ -31,8 +31,6 @@ export function resolveWhatsAppOutboundTarget(params: {
     if (isWhatsAppGroupJid(normalizedTo)) {
       return { ok: true, to: normalizedTo };
     }
-    // Enforce allowFrom for all direct-message send modes (including explicit).
-    // Group destinations are handled by group policy and are allowed above.
     if (hasWildcard || allowList.length === 0) {
       return { ok: true, to: normalizedTo };
     }
